docker/files/requirements.txt
factorio_rcon_py
requests